파일 매니저 컴포넌트 작업중

+++ client/components/file-manager/content.vue
@@ -64,8 +64,8 @@
           :mime="file.mime"
           :selected="!!file.selected"
           :badge="file.badge || []"
-          :context="[]"
-          @select="onSelect"/>
+          @select="onSelect"
+          @select-context-item="onSelectContextItem"/>
       </li>
     </ul>
   </div>
@@ -81,6 +81,8 @@ import { upload } from '../../libs/api.js'
 import { request } from '../../libs/api.js'
 import { convertDataToFileItem } from './libs.js'
 import { printf } from '../../libs/strings.js'
+import { sleep } from '../../libs/util.js'
+import { fileContextKey } from './assets.js'
 import { ButtonBasic, ButtonGroup } from '../button/index.js'
 import { Dropdown, Context } from '../navigation/dropdown/index.js'
 import Loading from '../content/loading.vue'
@@ -88,11 +90,13 @@ import Empty from '../content/empty.vue'
 import { File, Progress } from './item/index.js'
 
 const fileManager = inject('file-manager')
+const fileManagerEvent = inject('file-manager-event')
 const error = inject('error')
 const auth = authStore()
 const state = reactive({
   loading: false,
 })
+const errorPath = [ 'components', 'file-manager', 'content.vue' ]
 
 const _files = computed(() => {
   return fileManager.index
@@ -123,13 +127,15 @@ onMounted(async () => {
     })
     if (res?.data?.index?.length > 0)
     {
-      res.data.index.forEach(item => fileManager.addFile(convertDataToFileItem(item)))
+      res.data.index
+        .reverse()
+        .forEach(item => fileManager.addFile(convertDataToFileItem(item)))
     }
   }
   catch (e)
   {
     error.catch({
-      path: [ 'components', 'file-manager', 'content.vue', 'onMounted' ],
+      path: [ ...errorPath, 'onMounted' ],
       message: '첨부파일 목록을 가져오지 못했습니다.',
       error: e,
     })
@@ -164,11 +170,7 @@ async function uploadFiles()
   }
   catch (e)
   {
-    error.catch({
-      path: [ 'components', 'file-manager', 'content.vue', 'uploadFiles()' ],
-      message: typeof e === 'string' ? e : '파일 업로드하지 못했습니다.',
-      error: typeof e === 'string' ? new Error(e) : e,
-    })
+    errorUploadFile([ 'uploadFiles()' ], e)
   }
 }
 function uploadFile(files)
@@ -177,7 +179,13 @@ function uploadFile(files)
   let idx = undefined
   function nextFile(remainFiles, resolve)
   {
-    if (remainFiles.length > 0) uploadFile(remainFiles).then()
+    if (remainFiles.length > 0)
+    {
+      setTimeout(() => {
+        uploadFile(remainFiles)
+          .catch(e => errorUploadFile([ 'uploadFile()' ], e))
+      }, 100)
+    }
     else resolve()
   }
   return new Promise((resolve, reject) => {
@@ -188,7 +196,7 @@ function uploadFile(files)
       return nextFile(remainingFiles, resolve)
     }
     // checking limit file count
-    if (fileManager._countItems > fileManager.preference.limitCount)
+    if (fileManager._countItems >= fileManager.preference.limitCount)
     {
       return reject(printf('업로드할 수 있는 파일 갯수는 {0}개 입니다.', fileManager.preference.limitCount))
     }
@@ -228,6 +236,14 @@ function uploadFile(files)
     })
   })
 }
+function errorUploadFile(path, err)
+{
+  error.catch({
+    path: [ ...errorPath, ...path ],
+    message: typeof err === 'string' ? err : '파일 업로드 실패',
+    error: typeof err === 'string' ? new Error(err) : err,
+  })
+}
 
 function onSelect(idx, $event)
 {
@@ -247,29 +263,66 @@ async function onClickDelete()
   if (!confirm('선택한 파일을 삭제할까요? 삭제하면 복구할 수 없습니다.')) return
   const index = fileManager._selectedFilesIndex
   fileManager.selectAllFiles(false)
-  for (const key of index)
+  const _index = [ ...index ].reverse()
+  for (const key of _index)
   {
-    if (fileManager.items[key])
-    {
-      await deleteFile(fileManager.items[key].srl)
-      fileManager.removeFile(key)
-    }
+    if (!fileManager.items[key]) continue
+    await deleteFile(key)
+    await sleep(1000)
   }
 }
-async function deleteFile(srl)
+async function deleteFile(key)
 {
   try
   {
+    const srl = fileManager.items[key]?.srl
+    if (!srl) throw new Error('Not found srl.')
+    fileManager.removeFile(key)
     await request(`/file/${srl}/`, { method: 'delete' })
   }
   catch (e)
   {
     error.catch({
-      path: [ 'components', 'file-manager', 'content.vue', 'deleteFile()' ],
+      path: [ ...errorPath, 'deleteFile()' ],
       message: '파일을 삭제하지 못했습니다.',
       error: e,
     })
   }
+}
+
+async function onSelectContextItem(idx, code)
+{
+  const item = fileManager.items[idx]
+  if (!item)
+  {
+    error.catch({
+      path: [ ...errorPath, 'onSelectContextItem()' ],
+      message: '기능을 실행할 아이템이 없습니다.',
+    })
+    return
+  }
+  console.log('onSelectContextItem()', idx, code)
+  switch (code)
+  {
+    case fileContextKey.OPEN_NEW_WINDOW:
+      break
+    case fileContextKey.DOWNLOAD:
+      break
+    case fileContextKey.SET_THUMBNAIL:
+      break
+    case fileContextKey.INSERT_MARKDOWN:
+      break
+    case fileContextKey.INSERT_ADDRESS:
+      break
+    case fileContextKey.INSERT_HTML:
+      break
+    case fileContextKey.DELETE:
+      if (!confirm('선택한 파일을 삭제할까요? 삭제하면 복구할 수 없습니다.')) return
+      await deleteFile(idx)
+      break
+  }
+  // console.log('onSelectContextItem()', item)
+  // fileManagerEvent.
 }
 </script>
